requirement.txt
streamlit==1.31.1
requests==2.31.0
python-dotenv==1.0.0
pinecone-client==2.2.4
langchain==0.1.0
langchain-pinecone==0.1.1
langchain-openai==0.0.2
openai==1.4.0
PyMuPDF==1.23.0
tiktoken==0.5.1